Fixing the self request issue
